Episode 13

+++ kernel/src/math.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct Point{
-    unsigned int X;
-    unsigned int Y;
+    long X;
+    long Y;
 };
